feat: enforce --rpc.txfeecap flag (#15618)

+++ crates/ethereum/node/src/node.rs
@@ -362,7 +362,7 @@ where
             .with_head_timestamp(ctx.head().timestamp)
             .kzg_settings(ctx.kzg_settings()?)
             .with_local_transactions_config(pool_config.local_transactions_config.clone())
-            .set_tx_fee_cap(u64::try_from(ctx.config().rpc.rpc_tx_fee_cap)?)
+            .set_tx_fee_cap(u128::try_from(ctx.config().rpc.rpc_tx_fee_cap)?)
             .with_additional_tasks(ctx.config().txpool.additional_validation_tasks)
             .build_with_tasks(ctx.task_executor().clone(), blob_store.clone());
